Move type to shared module to avoid dependency from shared to sidebar store module
